add tests for getMessageCalldataCostUSD

+++ contracts/src/v0.8/ccip/test/priceRegistry/PriceRegistry.t.sol
@@ -7,8 +7,8 @@ import {RouterSetup} from "../router/RouterSetup.t.sol";
 import {PriceRegistry} from "../../PriceRegistry.sol";
 
 contract PriceRegistrySetup is TokenSetup, RouterSetup {
-  uint224 internal constant USD_PER_GAS = 1e6; // 0.001 gwei
-  uint224 internal constant USD_PER_CALLDATA_GAS = 1e9; // 1 gwei
+  uint112 internal constant USD_PER_GAS = 1e6; // 0.001 gwei
+  uint112 internal constant USD_PER_CALLDATA_GAS = 1e9; // 1 gwei
 
   // Encode L1 gas price of 1 gwei, and L2 gas price of 0.001 gwei
   uint224 internal constant PACKED_USD_PER_GAS = USD_PER_CALLDATA_GAS << 112 + USD_PER_GAS;
